backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector database client (for Qdrant)
qdrant-client>=1.7.0

# Database client (for Neon Postgres)
asyncpg>=0.29.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0

# Embeddings and LLM integration
cohere>=5.5.3

# For text processing and similarity search
numpy>=1.21.0
scikit-learn>=1.3.0

# For document processing
pypdf>=3.17.0
docx2txt>=0.8
python-pptx>=0.6.23

# Additional utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0